# tacocloud/tacocloud-api/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: TacoCloud course API
  version: 1.0.0
  description: 'TC-01 to TC-36. Money: MXN, sum of ingredient prices, HALF_UP at 2
    decimals. /api is a transitional alias with Deprecation: true and Link to this
    spec; no removal date is committed. Use /api/v1. Catalog metadata is academic
    and does not replace cross-contamination controls. Runtime Spring Boot 2.5 / MVC.'
servers:
- url: /api/v1
- url: /api
  description: Deprecated compatibility alias
paths:
...
  /auth/register:
    post:
      operationId: post_auth_register
      summary: Register /auth/register
      security: []
      parameters:
      - name: X-Correlation-Id
        in: header
        required: false
        schema:
          type: string
          pattern: ^[A-Za-z0-9_-]{1,64}$
        description: Invalid or absent values are replaced by a UUID.
      - name: X-XSRF-TOKEN
        in: header
        required: true
        schema:
          type: string
        description: Obtain from /auth/csrf and send with its XSRF-TOKEN cookie. Required
          with Basic and session authentication.
      responses:
        '201':
          description: Success
          headers:
            X-Correlation-Id:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '401':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '403':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '404':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '409':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '422':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '500':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrationRequest'
...
  /auth/login:
    post:
      operationId: post_auth_login
      summary: Sign in
      security: []
      parameters:
      - name: X-Correlation-Id
        in: header
        required: false
        schema:
          type: string
          pattern: ^[A-Za-z0-9_-]{1,64}$
        description: Invalid or absent values are replaced by a UUID.
      - name: X-XSRF-TOKEN
        in: header
        required: true
        schema:
          type: string
        description: Obtain from /auth/csrf and send with its XSRF-TOKEN cookie. Required
          with Basic and session authentication.
      responses:
        '200':
          description: Success
          headers:
            X-Correlation-Id:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '401':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '403':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '404':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '409':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '422':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '500':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
  /auth/me:
    get:
      operationId: get_auth_me
      summary: Current session
      security:
      - basicAuth: []
      - sessionAuth: []
      parameters:
      - name: X-Correlation-Id
        in: header
        required: false
        schema:
          type: string
          pattern: ^[A-Za-z0-9_-]{1,64}$
        description: Invalid or absent values are replaced by a UUID.
      responses:
        '200':
          description: Success
          headers:
            X-Correlation-Id:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '401':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '403':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '404':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '409':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '422':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '500':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
  /auth/csrf:
    get:
      operationId: get_auth_csrf
      summary: Initialize CSRF cookie
      security: []
      parameters:
      - name: X-Correlation-Id
        in: header
        required: false
        schema:
          type: string
          pattern: ^[A-Za-z0-9_-]{1,64}$
        description: Invalid or absent values are replaced by a UUID.
      responses:
        '200':
          description: Success
          headers:
            X-Correlation-Id:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CsrfResponse'
        '400':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '401':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '403':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '404':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '409':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '422':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '500':
          description: Controlled error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
...
components:
  schemas:
    IngredientType:
      type: string
      enum:
      - WRAP
      - PROTEIN
      - VEGGIES
      - CHEESE
      - SAUCE
    DietaryTag:
      type: string
      enum:
      - VEGAN
      - VEGETARIAN
      - GLUTEN_FREE
    Allergen:
      type: string
      enum:
      - GLUTEN
      - MILK
      - SOY
      - PEANUT
      - TREE_NUT
      - EGG
      - FISH
      - SHELLFISH
    Status:
      type: string
      enum:
      - CREATED
      - ACCEPTED
      - PREPARING
      - READY
      - OUT_FOR_DELIVERY
      - DELIVERED
      - CANCELLED
...
    IngredientResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          $ref: '#/components/schemas/IngredientType'
        unitPrice:
          type: number
          minimum: 0
          multipleOf: 0.01
          example: 7.0
        available:
          type: boolean
        dietaryTags:
          type: array
          items:
            $ref: '#/components/schemas/DietaryTag'
        allergens:
          type: array
          items:
            $ref: '#/components/schemas/Allergen'
        spiceLevel:
          type: integer
          minimum: 0
          maximum: 5
      additionalProperties: false
      required:
      - id
      - name
      - type
      - unitPrice
      - available
      - dietaryTags
      - allergens
      - spiceLevel
...
    TacoResponse:
      type: object
      properties:
        id:
          type: string
          nullable: true
        name:
          type: string
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/IngredientResponse'
        beverage:
          type: boolean
        dietaryTags:
          type: array
          items:
            $ref: '#/components/schemas/DietaryTag'
        allergens:
          type: array
          items:
            $ref: '#/components/schemas/Allergen'
        spiceLevel:
          type: integer
          minimum: 0
          maximum: 5
      additionalProperties: false
      required:
      - name
      - ingredients
      - beverage
      - dietaryTags
      - allergens
      - spiceLevel
...
    OrderLine:
      type: object
      properties:
        taco:
          $ref: '#/components/schemas/TacoResponse'
        quantity:
          type: integer
          minimum: 1
        unitPriceAtPurchase:
          type: number
          minimum: 0
          multipleOf: 0.01
          example: 7.0
        subtotal:
          type: number
          minimum: 0
          multipleOf: 0.01
          example: 7.0
      additionalProperties: false
      required:
      - taco
      - quantity
      - unitPriceAtPurchase
      - subtotal
    OrderResponse:
      type: object
      properties:
        deliveryName:
          type: string
          minLength: 1
          maxLength: 100
          nullable: true
        deliveryStreet:
          type: string
          minLength: 1
          maxLength: 200
          nullable: true
        deliveryCity:
          type: string
          minLength: 1
          maxLength: 100
          nullable: true
        deliveryState:
          type: string
          minLength: 1
          maxLength: 100
          nullable: true
        deliveryZip:
          type: string
          pattern: ^[A-Za-z0-9 -]{3,12}$
          nullable: true
        id:
          type: string
          nullable: true
        placedAt:
          type: string
          format: date-time
          nullable: true
        status:
          $ref: '#/components/schemas/Status'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderLine'
        subtotal:
          type: number
          minimum: 0
          multipleOf: 0.01
          example: 7.0
        discount:
          type: number
          minimum: 0
          multipleOf: 0.01
          example: 7.0
        total:
          type: number
          minimum: 0
          multipleOf: 0.01
          example: 7.0
        currency:
          type: string
          enum:
          - MXN
        version:
          type: integer
          nullable: true
        paymentBrand:
          type: string
          nullable: true
        paymentLast4:
          type: string
          pattern: ^[0-9]{4}$
          nullable: true
      additionalProperties: false
      required:
      - status
      - items
      - subtotal
      - discount
      - total
      - currency
...
    RegistrationRequest:
      type: object
      properties:
        username:
          type: string
          pattern: ^[a-zA-Z0-9_.-]{3,40}$
        password:
          type: string
          minLength: 10
          maxLength: 72
          writeOnly: true
        email:
          type: string
          format: email
        fullname:
          type: string
          nullable: true
        street:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        zip:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
      additionalProperties: false
      required:
      - username
      - password
      - email
    AccountResponse:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
      additionalProperties: false
      required:
      - id
      - username
      - email
    LoginRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
          writeOnly: true
      additionalProperties: false
      required:
      - username
      - password
    SessionResponse:
      type: object
      properties:
        username:
          type: string
        roles:
          type: array
          items:
            type: string
      additionalProperties: false
      required:
      - username
      - roles
    CsrfResponse:
      type: object
      properties:
        token:
          type: string
        headerName:
          type: string
          enum:
          - X-XSRF-TOKEN
      additionalProperties: false
      required:
      - token
      - headerName
...
    Violation:
      type: object
      properties:
        field:
          type: string
        code:
          type: string
        reason:
          type: string
      additionalProperties: false
      required:
      - field
      - reason
    ApiProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
        correlationId:
          type: string
        violations:
          type: array
          items:
            $ref: '#/components/schemas/Violation'
        previousTotal:
          type: number
          minimum: 0
          multipleOf: 0.01
          example: 7.0
        quote:
          $ref: '#/components/schemas/OrderResponse'
      additionalProperties: false
      required:
      - type
      - title
      - status
      - detail
      - instance
      - code
      - correlationId
      - violations
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    sessionAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID